[88] 아이폰 테스트
